requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.40.0
cryptography==42.0.5
Django==5.0.4
django-ckeditor-5==0.2.12
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.2
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jwcrypto==1.5.6
Markdown==3.6
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
pycparser==2.22
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
toposort==1.10
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
